Preprocessing the data differently now.
The pipeline now preprocesses the data now different. Now it takes
another argument, the so called method. This parameter can be char,
syllable or word. And depending on the value you passed the data will be
splitted into either characters, syllables or words. It should be
mentioned that special characters which are [^a-zA-Z0-9] will be kept.